requirements.txt
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.5
imbalanced-learn==0.13.0
joblib==1.5.1
kiwisolver==1.4.8
llvmlite==0.44.0
matplotlib==3.10.3
numba==0.61.2
numpy==2.2.6
packaging==25.0
pandas==2.3.0
pillow==11.3.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.6.1
scipy==1.16.0
seaborn==0.13.2
shap==0.48.0
six==1.17.0
sklearn-compat==0.1.3
slicer==0.0.8
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
